+++ src/Providers.tsx
@@ -33,7 +33,8 @@ import { TooltipProvider } from 'contexts/Tooltip';
 import { TransferOptionsProvider } from 'contexts/TransferOptions';
 import { TxMetaProvider } from 'contexts/TxMeta';
 import { UIProvider } from 'contexts/UI';
-import { ValidatorsProvider } from 'contexts/Validators';
+import { ValidatorsProvider } from 'contexts/Validators/ValidatorEntries';
+import { FavoriteValidatorsProvider } from 'contexts/Validators/FavoriteValidators';
 import { withProviders } from 'library/Hooks';
 import { PayoutsProvider } from 'contexts/Payouts';
 import { OverlayProvider } from '@polkadot-cloud/react';
@@ -61,6 +62,7 @@ export const Providers = withProviders(
   ActivePoolsProvider,
   TransferOptionsProvider,
   ValidatorsProvider,
+  FavoriteValidatorsProvider,
   FastUnstakeProvider,
   PayoutsProvider,
   UIProvider,
