+++ src/library/ListItem/Labels/ParaValidator.tsx
@@ -5,7 +5,7 @@ import { faCubes } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { useTranslation } from 'react-i18next';
 import { useTooltip } from 'contexts/Tooltip';
-import { useValidators } from 'contexts/Validators';
+import { useValidators } from 'contexts/Validators/ValidatorEntries';
 import { TooltipTrigger } from 'library/ListItem/Wrappers';
 import type { ParaValidatorProps } from '../types';
