+++ src/pages/Nominate/Active/Nominations/index.tsx
@@ -10,12 +10,13 @@ import { useHelp } from 'contexts/Help';
 import { useActivePools } from 'contexts/Pools/ActivePools';
 import { useStaking } from 'contexts/Staking';
 import { useUi } from 'contexts/UI';
-import { useValidators } from 'contexts/Validators';
+import { useValidators } from 'contexts/Validators/ValidatorEntries';
 import { CardHeaderWrapper } from 'library/Card/Wrappers';
 import { useUnstaking } from 'library/Hooks/useUnstaking';
 import { ValidatorList } from 'library/ValidatorList';
 import type { MaybeAccount } from 'types';
 import { useOverlay } from '@polkadot-cloud/react/hooks';
+import { useFavoriteValidators } from 'contexts/Validators/FavoriteValidators';
 import { Wrapper } from './Wrapper';
 
 export const Nominations = ({
@@ -26,18 +27,15 @@ export const Nominations = ({
   nominator: MaybeAccount;
 }) => {
   const { t } = useTranslation('pages');
-  const { openModal } = useOverlay().modal;
-  const { inSetup } = useStaking();
   const { isSyncing } = useUi();
-  const { activeAccount, isReadOnlyAccount } = useConnect();
-  const { getAccountNominations } = useBonded();
-  const { isFastUnstaking } = useUnstaking();
-  const { nominated: stakeNominated, poolNominated } = useValidators();
   const { openHelp } = useHelp();
-  let { favoritesList } = useValidators();
-  if (favoritesList === null) {
-    favoritesList = [];
-  }
+  const { inSetup } = useStaking();
+  const { openModal } = useOverlay().modal;
+  const { isFastUnstaking } = useUnstaking();
+  const { getAccountNominations } = useBonded();
+  const { favoritesList } = useFavoriteValidators();
+  const { activeAccount, isReadOnlyAccount } = useConnect();
+  const { nominated: stakeNominated, poolNominated } = useValidators();
 
   const {
     poolNominations,
