+++ src/pages/Validators/Stats/AverageCommission.tsx
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import { useTranslation } from 'react-i18next';
-import { useValidators } from 'contexts/Validators';
+import { useValidators } from 'contexts/Validators/ValidatorEntries';
 import { Text } from 'library/StatBoxList/Text';
 
 export const AverageCommissionStat = () => {
